feat: delete media attachment
