+++ src/Components/ChatComponent/components/Media/index.js
@@ -1,0 +1,1 @@
+export * from "./Media";
