+++ src/Components/ChatComponent/components/Message/index.js
@@ -1,2 +1,3 @@
-export * from "./SendedMessage"
-export * from "./ReceivedMessage"
+export * from "./SendedMessage";
+export * from "./ReceivedMessage";
+export * from "../Media";
